+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 Features:
 
-  - Perform searches on indexed classes to restrict the results to objects of a specific class
-  - Allow to specify multiple blueprint attributes and index methods in one statement (without specifying options)
+  - perform searches on indexed classes to scope the search to objects of a specific class
+  - specify multiple blueprint attributes and index methods in one statement (without specifying options)
   - use blocks for complex attribute or index specifications
 
 ##0.3.3 (December 13th, 2010)