+++ CHANGELOG.md
@@ -5,6 +5,9 @@ Features:
   - perform searches on indexed classes to scope the search to objects of a specific class
   - specify multiple blueprint attributes and index methods in one statement (without specifying options)
   - use blocks for complex attribute or index specifications
+  - changed the implementation of Resultset.size to get more accurate estimations
+  - changed the indexing of active_record or datamapper models when declared as attributes or indexes
+    in a blueprint (indexes now all attributes of the object instead of using to_s)
 
 ##0.3.3 (December 13th, 2010)